requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
mcp[cli]>=1.9.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
httpx>=0.28.0
openbb>=4.5.0
apscheduler>=3.11.0
cryptography>=44.0.0
python-jose>=3.3.0
pandas>=2.2.0
numpy>=1.26.0
